Move the helper functions to their own module
