+++ crates/sui-json-rpc-types/src/utils.rs
@@ -1,0 +1,29 @@
+use fastcrypto::encoding::Base64;
+use serde_json::{json, Value};
+
+/// Convert a json array of bytes to Base64
+pub(crate) fn bytes_array_to_base64(v: &mut Value) {
+    match v {
+        Value::Null | Value::Bool(_) | Value::Number(_) | Value::String(_) => (),
+        Value::Array(vals) => {
+            if let Some(vals) = vals.iter().map(try_into_byte).collect::<Option<Vec<_>>>() {
+                *v = json!(Base64::from_bytes(&vals).encoded())
+            } else {
+                for val in vals {
+                    bytes_array_to_base64(val)
+                }
+            }
+        }
+        Value::Object(map) => {
+            for val in map.values_mut() {
+                bytes_array_to_base64(val)
+            }
+        }
+    }
+}
+
+/// Try to convert a json Value object into an u8.
+pub(crate) fn try_into_byte(v: &Value) -> Option<u8> {
+    let num = v.as_u64()?;
+    (num <= 255).then_some(num as u8)
+}
